requirements.txt
litellm
pyautogen==0.2.31
qdrant_client==1.9.2
streamlit
pydantic
IPython
chromadb==0.5.3
pypdf
fastapi
fastembed==0.2.7
pyyaml
openpyxl
pydantic_settings
dash
dash-bootstrap-components
requests
sentence_transformers
logger
boto3==1.34.136
awscli==1.33.18
pyautogen[retrievechat]
litellm[proxy]
qdrant_client[fastembed]